requirements/game/app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bidict==0.23.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
frozenlist==1.4.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
multidict==6.1.0
pillow==10.4.0
propcache==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.32.3
service-identity==24.1.0
simple-websocket==1.0.0
djangorestframework==3.15.2
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
wsproto==1.2.0
yarl==1.14.0
zope.interface==7.0.3
neat-python==0.92
